requirements.txt
alembic==1.15.2
ansible==10.7.0
ansible-core==2.17.11
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
exceptiongroup==1.2.2
flasgger==0.9.5
Flask==2.2.5
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
fonttools==4.58.1
greenlet==3.2.0
idna==3.10
iniconfig==2.1.0
ipaddress==1.0.23
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lark==1.2.2
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mistune==3.1.3
networkx==3.4.2
numpy
packaging
pillow==11.2.1
pluggy==1.5.0
pycparser==2.22
pyparsing==3.2.3
pytest==7.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-hcl2==7.2.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
resolvelib==1.0.1
rpds-py==0.24.0
six==1.17.0
SQLAlchemy==2.0.25
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==2.3.8
langchain>=0.2.0
langchain-community
langchain-openai
openai
chromadb
unstructured[local-inference]
openstacksdk
celery==5.4.0
redis==5.0.7
PyJWT==2.8.0
gunicorn==22.0.0
PyMySQL==1.1.1